# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /studio/workspace:
    get:
      operationId: getStudioWorkspace
      tags: [studio]
      summary: Read the signed-in user's workspace
      responses:
        "200":
          description: The authenticated user's workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioWorkspaceResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No workspace exists for this account yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      operationId: saveStudioWorkspace
      tags: [studio]
      summary: Save the signed-in user's workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudioWorkspaceInput"
      responses:
        "200":
          description: The saved workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioWorkspaceResponse"
        "400":
          description: Invalid workspace data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: The workspace changed since it was loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioWorkspaceConflict"
  /studio/legal-profile:
    get:
      operationId: getStudioLegalProfile
      tags: [studio]
      summary: Get the public legal profile shared by this Remix
      responses:
        "200":
          description: Public company details and whether the signed-in account can manage them
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioLegalProfileResponse"
    put:
      operationId: saveStudioLegalProfile
      tags: [studio]
      summary: Save the Remix-wide legal profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudioLegalProfileInput"
      responses:
        "200":
          description: The saved public company details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioLegalProfileResponse"
        "400":
          description: Invalid company details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Another account manages this Remix's legal profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    StudioWorkspaceState:
      type: object
      additionalProperties: true
    StudioWorkspaceInput:
      type: object
      additionalProperties: false
      properties:
        state:
          $ref: "#/components/schemas/StudioWorkspaceState"
        expectedVersion:
          type: ["integer", "null"]
          minimum: 1
      required:
        - state
        - expectedVersion
    StudioWorkspaceResponse:
      type: object
      additionalProperties: false
      properties:
        state:
          $ref: "#/components/schemas/StudioWorkspaceState"
        version:
          type: integer
          minimum: 1
      required:
        - state
        - version
    StudioWorkspaceConflict:
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
        current:
          $ref: "#/components/schemas/StudioWorkspaceResponse"
      required: [error, current]
    StudioLegalProfileInput:
      type: object
      additionalProperties: false
      properties:
        registeredName:
          type: string
          maxLength: 200
        tradingName:
          type: string
          maxLength: 200
        country:
          type: string
          maxLength: 120
        registeredAddress:
          type: string
          maxLength: 1000
        privacyEmail:
          type: string
          maxLength: 320
        website:
          type: string
          maxLength: 500
      required:
        - registeredName
        - tradingName
        - country
        - registeredAddress
        - privacyEmail
        - website
    StudioLegalProfileResponse:
      type: object
      additionalProperties: false
      properties:
        registeredName:
          type: string
        tradingName:
          type: string
        country:
          type: string
        registeredAddress:
          type: string
        privacyEmail:
          type: string
        website:
          type: string
        updatedAt:
          type: ["string", "null"]
          format: date-time
        isComplete:
          type: boolean
        canEdit:
          type: boolean
      required:
        - registeredName
        - tradingName
        - country
        - registeredAddress
        - privacyEmail
        - website
        - updatedAt
        - isComplete
        - canEdit
...
    ApiError:
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
      required:
        - error