# lib/api-zod/src/generated/types/studioLegalProfileResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudioLegalProfileResponse {
  registeredName: string;
  tradingName: string;
  country: string;
  registeredAddress: string;
  privacyEmail: string;
  website: string;
  /** @nullable */
  updatedAt: Date | null;
  isComplete: boolean;
  canEdit: boolean;
}