# lib/api-zod/src/generated/types/studioWorkspaceConflict.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudioWorkspaceResponse } from './studioWorkspaceResponse';

export interface StudioWorkspaceConflict {
  error: string;
  current: StudioWorkspaceResponse;
}